Make media-session its own standalone package
This removes the rest of pipewire from the tree, but keeps the git repo.
The package name is now just 'media-session' but the binary remains
pipewire-media-session for backwards compatibility.

No functional changes besides the removal but a few renames for
variables from PIPEWIRE_FOO to MEDIA_SESSION_FOO (and its lowercase
version).

In docs/ folder, the media-session.dox moved to be the new mainpage
index.dox and all other dox files are removed.

pw-uninstalled.sh becomes media-session-uninstalled.

Finally, the hierarchy was flattened, so
- src/media-session/* goes into src/
- src/daemon/systemd goes into systemd/
- src/media-session/media-session.d goes into media-session.d